fix error reporting of runner
When an error is reported, container.logs() returns an iterator over
lines of bytes. The decode() function should apply to each line, not to
the iterator, otherwise we get no logs, but an error message saying
there is no decode() function in the iterator.

This change follows the same pattern as stream_logs() in the same file.